Move out model query logic from dialog and fix imports
Fetch chat string from tuple for parsing

fix greetings
