+++ droidlet/perception/semantic_parsing_model/safety_words.py
@@ -1,0 +1,2 @@
+
+
