+++ droidlet/perception/semantic_parsing_model/utils/preprocess.py
